Add `kart lfs+ gc` - cleans up unreferenced LFS tiles
Very bare-bones - has a name which only makes sense to those who
understand our underlying technology, and simply deletes anything that
is not currently in use.

Still TODO, later:
A bit more design on how we expose this - what to call it, how to make
sure users know about it.
Maybe more options for automatic garbage collection, or, inform the user
once they are wasting a certain amount of space...