Improve understanding of revisions vs filters [#706]
Don't assume things are filters just because they seem not to be
revisions - they may be a revision with a typo in them. Instead,
check to make sure filters are present at HEAD, and fail with
an error message if we can't decide if something is a revision or
a filter.

Also fixes up some naming.