add cffi as a wheel dependency under vcpkg-vendor
cffi depends on libffi — this is part of the vcpkg ecosystem already,
so lets use it rather than pick up a system library.

Note until this is treated the same as pygit/etc in `make py-reqs`
we're relying on pip not installing the version from requirements.txt
because we've already installed the wheel. Version updates will break
this behaviour.