Fix PyCreateVirtualEnvironment to work on windows
Using cmake -E env --modify, removes the need to have any variables that
contain second-order semicolons - ie a semicolon that must not be a list
delimiter. If the variables did contain such semicolons, they can quickly
go wrong here in PyCreateVirtualEnvironment but also in the parent context
when it uses variables likes ${TARGET}_PYTHON - which is pretty untenable.
Every time this value is used or copied into another variable is another
possibility for very non-obvious bug to occur.

Downside: requires cmake 3.25 which is cutting edge.