Fix gdal to build on windows - gdal-config
Don't copy_tools for gdal-config on windows since that looks for
gdal-config.exe
gdal-config is a shell script and so is not an exe even on windows.
On other platforms, copy_tools works for executables AND shell scripts.

Basic fix, this means gdal-config shell script is simply missing on
windows but this fixes the build - remains to be seens if a better fix
is needed.