Accept but don't require nested-archives
Github zips up CI artifacts, so we end up with
(vendor-Windows.zip).zip
(vendor-Darwin.tar.gz).zip
(vendor-Linux.tar.gz).zip

The actual filenames of these are simply:
vendor-Windows.zip
vendor-Darwin.zip
vendor-Linux.zip

This change lets the user supply either the inner or outer archive
as -DVENDOR_ARCHIVE - we just unarchive it twice if needed.
Note that the name of the archive is not necessarily sufficient to tell
us if it is the inner or outer archive.