Add logging.config into pyinstaller packages
If ext-run code wants to configure logging properly it is likely
to want `logging.config.dictConfig()`, and this is not currently
included in Kart. It's also hard to include via sys.path hackery since
`logging` *is* included in Kart.

This change just adds all sub-modules of `logging` into the pyinstaller
packages.