Remove special-casing for windows event loop
This code was once needed to avoid a crash on windows due to
using the asyncio.SelectorEventLoop.
From https://docs.python.org/3.6/library/asyncio-eventloops.html#asyncio.SelectorEventLoop
On Windows, only sockets are supported (ex: pipes are not supported).

Details of the crash are now lost but could be rediscovered by
re-testing commit f7d2d50

This workaround is apparently no longer necessary since python3.8
where the ProactorEventLoop is now the default on Windows.

And, this workaround now is causing errors of its own:
RuntimeError: There is no current event loop in thread 'MainThread'

Probably this is because this is not being called on the main thread -
it is an error to call get_event_loop if there is no event loop
AND the current thread is not the main thread.
But in any case, this workaround is not needed, calling get_event_loop
is deprecated, and asyncio.run now does everything we need.