Help the user fix their working copy if it gets into a bad state
If Kart crashes or otherwise is aborted partway through a checkout operation,
it is possible for the contents of the working copy or individual working copy
parts to end up out of sync with what should be checked out based on the git
file `.kart/HEAD`.

Until now, this caused an internal error to be raised - WorkingCopyTreeMismatch,
which gave some clues as to what had gone wrong but no real help on how to fix it.
Since Kart checks the WC state is valid as a precondition to most commands, arriving
in this state with no clear idea of how to get out of it makes a Kart repo feel
pretty broken, as most commands will simply raise the same internal error.

This change adds an external error including error code (which also suppresses the
stack trace). The error message contains instructions on how to fix the repo so
that it can be used again, and if run in an interactive terminal, prompts
the user to decide if they want to try this remedy immediately.