macos: rationalise macosx-deployment-target
We default to something sensible for cmake/vcpkg building: 10.15
on amd64 and 11.0 on arm64. But if the python interpreter was built
with greater than that we need to use the highest one for wheel building,
or we get clashes with wheel naming, since some builds figure out
there's a difference.

So detect that at configure time and issue a warning if there's a
difference.