Refactor: modify schema class to be more generic.
Tabular schemas look something like this:
[{
  "id": "deadbeef",
  "name": "fid",
  "dataType": "integer",
  "primaryKeyIndex": 0,
  "size": 64}, ...
]

Point cloud schemas look something like this:
[{"name": "Red", "dataType": "integer", "size": 32}, ...]

Raster schemas will look different again.

Removing the requirement that schemas themselves follow a particular schema,
and, doing away with the two different internal representations of schemas:
the list of dicts (json-dumpable) and the Schema object (not json-dumpable).
With the magic of python, it's possible to make the Schema object *be* a
list of dicts (actually a tuple) and be JSON dumpable, so, we don't need
to convert between the two all the time.