Basic raster import
Re-use tile import code from point-cloud -
plus some new metadata extraction code using gdalinfo -
to get the import of a GeoTIFF working, and add a unit test.

Any kind of conversion during import (ie, convert-to-cog) is not supported.
Extraction of color-tables / category-tables (which should be part of the schema)
is not yet supported.
Detection of COG vs non-COG is not yet supported.
Exact storage of envelopes, file-format etc is still up in the air.
The file-system WC has not yet been modified to check out this type of dataset.