Commit raster tiles and PAM files
- Diffs to raster tiles can now be committed
- Diffs to PAM files can now be committed

Attempts to deal with the possibility that the
PAM file's case doesn't match up with the raster tile,
at least at commit time - tries to normalise everything as follows:

- the raster tilename can be anything (upper or lowercase)
- internally, the raster tile extension is not stored
- externally, the raster tile extension is ".tif" in the WC
  (we could make this configurable to .tiff if it would help anyone)
- internally, the PAM filename is the same as the tilename, + ".aux.xml"
- externally the WC PAM name is the same as the tile + ".aux.xml" in the WC

but, we have to allow for deviations from this at commit-time and fix
them up during the commit.

Also: renamed meta/bands to meta/band - we don't use plurals
for any of the other dataset trees.
Some adjustment to how pam deltas are generated from previous PR.