Bugfix - don't require `user.name` and `user.email`...
when GIT_AUTHOR_NAME, GIT_AUTHOR_EMAIL, GIT_COMMITTER_NAME and
GIT_COMMITTER_EMAIL are set in the environment.

There's some complexity here that makes it tricky to just check we have
exactly what we need: Kart never actually needs to know the user name
or user email. What it needs to know are the author or committer's name
and email. And in practise it always does know these, since, the OS
will probably be able to supply one (eg by `echo $USER`), and these
defaults from the OS are picked up by the calls we use:
`git var GIT_AUTHOR_IDENT` and `git var GIT_COMMITTER_IDENT`

The simplest fix that keeps the good behaviour we have - prompting new
users to fill in user name and email if they haven't already - without
messing things up for power users who want to set those env vars - is just
to keep the check as is, but disable it if any of the env vars are set.