Support convert-to-cog during import
And make it the default. Doesn't yet work during diffs / commits.

Updated format.json in line with point clouds - here is a PC format.json:
{
  "compression": "laz",
  "lasVersion": "1.4",
  "optimization": "copc",
  "optimizationVersion": "1.0",
  "pointDataRecordFormat": 7,
  "pointDataRecordLength": 36
}

And here is the current raster format.json:
{
  "fileType": "image/tiff; application=geotiff; profile=cloud-optimized"
}

Here is the new raster format.json:
{
  "fileType": "image/tiff; application=geotiff",
  "profile": "cloud-optimized"
}

Consistent with the PC format.json, each of these items represents a constraint:
1. the fileType must be geotiff
2. the profile must be "cloud-optimized"
One of these constraints can be removed to make a dataset which
allows geotiffs of any profile, by omitting that item from the JSON.

This meant updating the test data again.