Switch amazonlinux:2 installtest to amazonlinux:2023
The latest sqlite on amazonlinux:2 is 10 years old, and is
incompatible with the mod_spatialite extension that we bundle with Kart.
This causes the e2e tests to fail since they try to use the system sqlite
package plus that extension to make edits to the working copy.

amazonlinux:2023 has a newer sqlite that we can use in the e2e tests.