Require user to supply --convert-to-dataset-format
... or, sometimes, --no-convert-to-dataset-format

--convert-to-dataset-format: the user acknowledges that the tiles
as they are don't comply to the dataset format constraint. Kart
is allowed to convert them during commit so that they do.
In practise, the conversions will be any of:
make cloud optimized
change laz version (eg LAZ 1.2 to LAZ 1.4)

--no-convert-to-dataset-format: the user acknowledges that the
tiles as they are don't comply to the dataset format constraint.
Kart is allowed to update the dataset format so that they do.
In practice the update will be:
- drop cloud optimized constraint
- change laz format of entire dataset (only allowed if every new tile conforms
  to the new format)