Fix tab-completion to work with helper mode
Click tab completion uses the name of the program for various
environment variables eg _KART_COMPLETE.
Our program name varies depending on how the program is run
kart vs kart_cli depending on helper mode etc.
Since we can't be sure if the way kart is currently being run is the
same as it was run when tab-completion was installed, its safest to
stick to one "canoncial" program name for tab-completion at all times.
Click lets us do this. The chosen name is "kart" and therefore the
variable is _KART_COMPLETE.