Fix tests that check no-crs behaviour
No-crs tests weren't doing a lot since the no-crs test repo was bare,
so, no working copy was being created.
Fixed the no-crs test repo to not be bare, and, modified
`kart create-workingcopy` so that it complains if you run it on a bare
repo, which would have caused these tests to fail.
Also, the tests then run `kart status` and checks for success,
which basically just checks the repo is not corrupt. A better test
is to make sure there are no spurious diffs on this freshly
checked-out dataset, eg using `kart diff --exit-code`