Adds an experimental byod-point-cloud-import command
Imports the metadata from some PC tiles hosted on S3 without
downloading them, except, right now it does download them, since
a) PDAL can't extract metadata from S3 without downloading
the whole tile anyway and
b) boto3 is now configured to fetch from S3 during testing, but
PDAL is not, and in the long run, we might not need it to anyway
(see TODO below).

Still TODO: use boto3 or similar to download certain ranges of a PC
file, which we can then extract metadata from, without downloading
all the points. (Or fix PDAL so that it can do this).