Don't use lru_cache for method memoization
Since the cache itself has per-class lifetime, and all arguments are
used (with strong refs) as part of the key, this means that memoizing
methods with lru_cache will easily cause memory leaks in long-lived
processes.

Instead, use cached_property since all of these things are properties
anyway. cached_property stores the cached result on the instance itself
so that it shares the instance lifecycle.