Fix CURL_CA_INFO incorrect when using kart helper
Kart sets CURL_CA_INFO at startup in __init__.py and this also happens
in helper mode.

Unfortunately, in helper mode the environment is overwritten with the
environment of the calling process, and CURL_CA_INFO disappears.

Then if the helper invokes PDAL or any other subprocess it doesn't get
that variable.

This fix adds CURL_CA_INFO to the whitelist of env vars to preserve in
the helper rather than override from the calling process.