Improve existing --no-checkout flag
kart import supports --no-checkout, but it merely postpones
creating a working copy. As of this commit, it will add
also add the dataset to a set of non-checkout datasets in the
config, so the dataset will stay out of the WC indefinitely.

Also adds `kart checkout --dataset=foo --not-dataset=bar` flags
to add and remove datasets from the set of non-checkout datasets