Added delta-filters
Backwards compatible in that they do nothing unless you specify
--delta-filter=<some-value>
A bit like minimal-patches which we implemented a couple years back,
but more powerful - more general-purpose - to avoid having to implement
them again at some point.
Basically, lets a consumer of JSON or JSON-LINES diff opt in or out
of any of the following:
- inserts
- updates (old values)
- updates (new values)
- deletes
But still see where the filtered out data is.

Useful for generating changeset views of the data that don't need to show old values.