Fix merge test on linux
the HOME override in the merge test was undoing a fix for the
`safe.directory` git config option which was set by the
`git_user_config` fixture.

This change switches to using a copy of that config file but with the
`user.name` and `user.email` removed, which fixes the broken test.