vendor: avoid py build-isolation for internal deps
`python -m build` uses an isolated venv by default and reinstalls build dependencies like cffi.

We already build cffi ourselves, and want to use that, so we need to avoid the isolation.

We have a mechanism for creating an isolated build venv, so use that instead for both pygit2 & reflink