Don't wrap feature deltas with RawDiffDelta
They're already fine as pygit2.DiffDelta instances.
This change adjusts the API of RawDiffDelta so it is compatible with
DiffDelta, so the types can be used interchangeably.

This change allows us to look up feature blobs from their OID, which is
known from the delta objects, saving a path lookup. This makes
diff TTFB slightly faster (~10% faster)