Add a LazyDeltaDiff class
This is a refactoring of #1032; the aim is to make the lazy-ness of the
DeltaDiff class more explicit.

A LazyDeltaDiff doesn't behave like a dict, unlike other diff structs.
It must be consumed either via the `iter_items()` method or by calling
`resolve()` to convert it to a `DeltaDiff` first.