Fix an error with reprojected diffs in non-text output
Non-text diff writers need to lookup CRS when doing reprojection,
and they do so by checking whether the CRS has changed in the diff.

When --no-sort-keys is given, it's possible for the meta deltas to be
lazy, and if so then the diff writer will fail to find the CRS in the
diff.

This change fixes the issue by making meta deltas always non-lazy, and
adds a regression test