Fix garbled JSON-lines output sometimes
In `kart diff -o json-lines --add-feature-count-estimate={value}`,
two threads would write to the output. This correctly used a lock to
prevent races, but there is an intermediate bytearray which is written
just beforehand (introduced with #1025) and this also needed to use
a lock to prevent races.

This change moves that intermediate bytearray under the same lock as
writing the output buffer, thus avoiding the race condition