merge: When `-m` is given, don't imply `--no-ff`
Previously, there was no way to specify a message while using the
default `--ff` behaviour. The message is useful in the case that the
command can't do a fastforward and has to fallback to a merge commit.

This change makes it so that, if kart does decide to use a merge commit,
it will use the commit message given.